Fix decimal point removed after first 0
Fixes an issue where the decimal point after
the first 0 on display would be removed after
entering the next digit (e.g instead of 0.3,
it would become 03). This was due to the if
statement which handles the replacement of the 0
number that would trigger when 0 is the only one
on display. Also, another issue that involved the
0 not appearing when display was empty was fixed