Add backspace keyboard support
Pressing the backspace button now allows the user
to delete the last digit, just like the backspace
button does. On the topic, a bug involving that
button, where it would output "backspace" when the
slice second argument was set to -1, was finally
fixed. In fact, this was due to the "backspace"
button being mistakenly included in the "digit"
class, now it's been replaced with the "backspace"
class