Handle user pressing same operator twice
Takes into consideration whenever the user presses
the operator twice after entering a first number,
and makes it so the second number is same as
first (e.g 5 ++ is like doing 5 + 5 = 10);
Also handles whenever user presser the operator
twice after entering the second number, it will
act similar to the equal button. However, these
won't work if user changes the operator after
selecting the first one (e.g first choose "+",
the chooses "*")Also in all instances, even when
equal is used, it will restore the
"firstNum", "secondNum" and "operatorType", so
that it can use the result value in
display.TextContent as the "firstNum"
for subsequent operations.